# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Hensun Desk Cloud",
    "description": "Commercial pilot control plane; not a production certification.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/account/entitlement": {
      "get": {
        "tags": [
          "subscriptions"
        ],
        "summary": "Entitlement",
        "operationId": "entitlement_v1_account_entitlement_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntitlementResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/account/usage": {
      "get": {
        "tags": [
          "subscriptions"
        ],
        "summary": "Usage",
        "operationId": "usage_v1_account_usage_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsageSummaryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EntitlementResponse": {
        "properties": {
          "plan": {
            "type": "string",
            "title": "Plan"
          },
          "monthly_turn_limit": {
            "type": "integer",
            "title": "Monthly Turn Limit"
          },
          "used_turns": {
            "type": "integer",
            "title": "Used Turns"
          },
          "remaining_turns": {
            "type": "integer",
            "title": "Remaining Turns"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          }
        },
        "type": "object",
        "required": [
          "plan",
          "monthly_turn_limit",
          "used_turns",
          "remaining_turns",
          "expires_at"
        ],
        "title": "EntitlementResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "UsageSummaryResponse": {
        "properties": {
          "voice_turns": {
            "type": "integer",
            "title": "Voice Turns"
          },
          "provider_cost_micros": {
            "type": "integer",
            "title": "Provider Cost Micros"
          },
          "pricing_configured": {
            "type": "boolean",
            "title": "Pricing Configured"
          },
          "asr_units": {
            "type": "integer",
            "title": "Asr Units"
          },
          "llm_input_units": {
            "type": "integer",
            "title": "Llm Input Units"
          },
          "llm_output_units": {
            "type": "integer",
            "title": "Llm Output Units"
          },
          "tts_units": {
            "type": "integer",
            "title": "Tts Units"
          }
        },
        "type": "object",
        "required": [
          "voice_turns",
          "provider_cost_micros",
          "pricing_configured",
          "asr_units",
          "llm_input_units",
          "llm_output_units",
          "tts_units"
        ],
        "title": "UsageSummaryResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}